minor refactoring

+++ src/qcommon/qcommon.h
@@ -1131,13 +1131,13 @@ temp file loading
 #define Z_TagMalloc(size, tag)          Z_TagMallocDebug(size, tag, # size, __FILE__, __LINE__)
 #define Z_Malloc(size)                  Z_MallocDebug(size, # size, __FILE__, __LINE__)
 #define S_Malloc(size)                  S_MallocDebug(size, # size, __FILE__, __LINE__)
-void *Z_TagMallocDebug(int size, int tag, char *label, char *file, int line);   // NOT 0 filled memory
-void *Z_MallocDebug(int size, char *label, char *file, int line);               // returns 0 filled memory
-void *S_MallocDebug(int size, char *label, char *file, int line);               // returns 0 filled memory
+void *Z_TagMallocDebug(size_t size, int tag, char *label, char *file, int line);   // NOT 0 filled memory
+void *Z_MallocDebug(size_t size, char *label, char *file, int line);               // returns 0 filled memory
+void *S_MallocDebug(size_t size, char *label, char *file, int line);               // returns 0 filled memory
 #else
-void *Z_TagMalloc(int size, int tag);   // NOT 0 filled memory
-void *Z_Malloc(int size);               // returns 0 filled memory
-void *S_Malloc(int size);               // NOT 0 filled memory only for small allocations
+void *Z_TagMalloc(size_t size, int tag);   // NOT 0 filled memory
+void *Z_Malloc(size_t size);               // returns 0 filled memory
+void *S_Malloc(size_t size);               // NOT 0 filled memory only for small allocations
 #endif
 void Z_Free(void *ptr);
 void Z_FreeTags(int tag);
@@ -1150,7 +1150,7 @@ qboolean Hunk_CheckMark(void);
 //void *Hunk_Alloc( int size );
 // void *Hunk_Alloc( int size, ha_pref preference );
 void Hunk_ClearTempMemory(void);
-void *Hunk_AllocateTempMemory(unsigned int size);
+void *Hunk_AllocateTempMemory(size_t size);
 void Hunk_FreeTempMemory(void *buf);
 int Hunk_MemoryRemaining(void);
 void Hunk_SmallLog(void);
